requirements.txt
requests>=2.31.0
pandas>=1.3.0
openpyxl>=3.0.0
pytest>=7.4.0
pytest-html>=4.1.1
python-dotenv>=1.0.0
loguru>=0.6.0
jinja2>=3.0.0
aiohttp>=3.8.0
python-json-logger>=2.0.7
yagmail>=0.15.0
pyyaml>=6.0.0 